base setup done
