+++ src/index.css
@@ -1,10 +1,16 @@
-/* ::-webkit-scrollbar{
-    width: 0px;
-    background: transparent;
-}
+
 html{
-  scrollbar-width: none;
-} */
+ scroll-behavior: smooth;
+} 
+::-webkit-scrollbar{
+  background-color: transparent;
+  width: 0.2rem;
+}
+::-webkit-scrollbar-thumb{
+  border-radius: 1rem;
+  background-color: red;
+}
+
 
 *{
   margin: 0;
@@ -15,6 +21,7 @@ box-sizing: border-box;
 svg{
   cursor: pointer;
 }
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
